refactor fhir converter container
